Replace jsmin with json5 dependency
THe jsmin library is meant for minification of Javascript, not loading
of JSON files with comments. It's also not likely to see much
development in future because the author has little interest in
supporting ES6 etc.

The json5 library (and the cython version pyjson5) are both actually
supported upstream, and more accurately fulfil the requirement of being
able to use comments in JSON files.